Move IExperiment to shared dir
