+++ src/mobile-app/components/experiment-picker.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { IExperiment, useExperiments } from "../hooks/use-experiments";
+import { useExperiments } from "../hooks/use-experiments";
+import { IExperiment } from "../../shared/experiment-schema-types";
 
 interface IProps {
   setExperiment: (experiment: IExperiment) => void;
