+++ src/shared/index.html
@@ -6,7 +6,6 @@
     <meta name="description" content="Mobile App">
     <link rel="shortcut icon" href="favicon.ico">
     <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
-<!--    <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/flatly/bootstrap.min.css" crossorigin="anonymous">-->
   </head>
   <body>
     <h2>Shared Components Demo Page</h2>
